mythlcdserver: rename commandlineparser.* to mythlcdserver_…
